PostgreSQL: migraciones adaptadas, enums OrderStatus/OrderItemStatus y mejoras

+++ prisma/migrations/20260212000000_add_package_price_to_test_profile/migration.sql
@@ -1,2 +1,2 @@
 -- AlterTable
-ALTER TABLE "TestProfile" ADD COLUMN "packagePrice" REAL;
+ALTER TABLE "TestProfile" ADD COLUMN "packagePrice" DOUBLE PRECISION;
